requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asttokens==2.4.1
    # via stack-data
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via nltk
comm==0.2.1
    # via ipywidgets
contourpy==1.1.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.1.1
    # via ipython
dnspython==2.4.2
    # via pymongo
executing==2.0.1
    # via stack-data
filelock==3.13.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.43.1
    # via matplotlib
fsspec==2023.10.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.17.3
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.4
    # via requests
ipython==8.20.0
    # via ipywidgets
ipywidgets==8.1.1
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jinja2==3.1.2
    # via torch
joblib==1.3.2
    # via
    #   nltk
    #   scikit-learn
jupyterlab-widgets==3.0.9
    # via ipywidgets
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.3
    # via jinja2
matplotlib==3.8.0
    # via -r requirements.in
matplotlib-inline==0.1.6
    # via ipython
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
nltk==3.8.1
    # via sentence-transformers
numpy==1.26.1
    # via
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   torchvision
    #   transformers
opencv-python==4.8.1.78
    # via -r requirements.in
packaging==23.2
    # via
    #   huggingface-hub
    #   matplotlib
    #   transformers
parso==0.8.3
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.1.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   torchvision
prompt-toolkit==3.0.43
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.17.2
    # via ipython
pymongo[srv]==4.5.0
    # via
    #   -r requirements.in
    #   pymongo
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via matplotlib
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
regex==2023.10.3
    # via
    #   nltk
    #   transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   torchvision
    #   transformers
safetensors==0.4.0
    # via transformers
scikit-learn==1.3.2
    # via sentence-transformers
scipy==1.11.3
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.2.2
    # via -r requirements.in
sentencepiece==0.1.99
    # via sentence-transformers
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
stack-data==0.6.3
    # via ipython
sympy==1.12
    # via torch
threadpoolctl==3.2.0
    # via scikit-learn
tokenizers==0.14.1
    # via transformers
torch==2.1.0
    # via
    #   sentence-transformers
    #   torchvision
torchvision==0.16.0
    # via sentence-transformers
tqdm==4.66.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   nltk
    #   sentence-transformers
    #   transformers
traitlets==5.14.1
    # via
    #   comm
    #   ipython
    #   ipywidgets
    #   matplotlib-inline
transformers==4.34.1
    # via sentence-transformers
typing-extensions==4.8.0
    # via
    #   huggingface-hub
    #   torch
urllib3==2.0.7
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
widgetsnbextension==4.0.9
    # via ipywidgets
python-dotenv